requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
openai
anthropic
groq
slowapi